Signature help test update
