+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test.Common/LanguageServerTestBase.cs
@@ -5,20 +5,13 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.LanguageServer;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
-using Microsoft.AspNetCore.Razor.LanguageServer.Test.Common;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.ExternalAccess.Razor;
 using Microsoft.CodeAnalysis.Razor;
-using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
-using Microsoft.VisualStudio.Composition;
-using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Moq;
 
 namespace Microsoft.AspNetCore.Razor.Test.Common
@@ -50,98 +43,6 @@ namespace Microsoft.AspNetCore.Razor.Test.Common
         internal FilePathNormalizer FilePathNormalizer { get; }
 
         protected ILoggerFactory LoggerFactory { get; }
-
-        protected static AdhocWorkspace CreateTestWorkspace(IEnumerable<(Uri documentUri, SourceText csharpSourceText)> files, ExportProvider exportProvider)
-        {
-            var workspace = TestWorkspace.Create() as AdhocWorkspace;
-
-            // Add project and solution to workspace
-            var projectInfo = ProjectInfo.Create(
-                id: ProjectId.CreateNewId("TestProject"),
-                version: VersionStamp.Default,
-                name: "TestProject",
-                assemblyName: "TestProject",
-                language: LanguageNames.CSharp,
-                filePath: "C:\\TestSolution\\TestProject.csproj");
-
-            var solutionInfo = SolutionInfo.Create(
-                id: SolutionId.CreateNewId("TestSolution"),
-                version: VersionStamp.Default,
-                projects: new ProjectInfo[] { projectInfo });
-
-            workspace.AddSolution(solutionInfo);
-
-            // Add document to workspace. We use an IVT method to create the DocumentInfo variable because there's a special constructor in Roslyn that will
-            // help identify the document as belonging to Razor.
-            var languageServerFactory = exportProvider.GetExportedValue<IRazorLanguageServerFactoryWrapper>();
-
-            var documentCount = 0;
-            foreach (var (documentUri, csharpSourceText) in files)
-            {
-                var documentFilePath = documentUri.AbsolutePath;
-                var textAndVersion = TextAndVersion.Create(csharpSourceText, VersionStamp.Default, documentFilePath);
-                var documentInfo = languageServerFactory.CreateDocumentInfo(
-                    id: DocumentId.CreateNewId(projectInfo.Id),
-                    name: "TestDocument" + documentCount,
-                    filePath: documentFilePath,
-                    loader: TextLoader.From(textAndVersion),
-                    razorDocumentServiceProvider: new TestRazorDocumentServiceProvider());
-
-                workspace.AddDocument(documentInfo);
-                documentCount++;
-            }
-
-            return workspace;
-        }
-
-        protected static async Task<CSharpTestLspServer> CreateCSharpLspServerAsync(AdhocWorkspace workspace, ExportProvider exportProvider, ServerCapabilities serverCapabilities)
-        {
-            var clientCapabilities = new ClientCapabilities();
-            var testLspServer = await CSharpTestLspServer.CreateAsync(workspace, exportProvider, clientCapabilities, serverCapabilities).ConfigureAwait(false);
-            return testLspServer;
-        }
-
-        private class TestRazorDocumentServiceProvider : IRazorDocumentServiceProvider
-        {
-            public bool CanApplyChange => throw new NotImplementedException();
-
-            public bool SupportDiagnostics => throw new NotImplementedException();
-
-            TService IRazorDocumentServiceProvider.GetService<TService>()
-            {
-                var serviceType = typeof(TService);
-
-                if (serviceType == typeof(IRazorSpanMappingService))
-                {
-                    return (TService)(IRazorSpanMappingService)new TestRazorSpanMappingService();
-                }
-
-                if (serviceType == typeof(IRazorDocumentPropertiesService))
-                {
-                    return (TService)(IRazorDocumentPropertiesService)new TestRazorDocumentPropertiesService();
-                }
-
-                return this as TService;
-            }
-
-            private class TestRazorSpanMappingService : IRazorSpanMappingService
-            {
-                public Task<ImmutableArray<RazorMappedSpanResult>> MapSpansAsync(
-                    Document document,
-                    IEnumerable<TextSpan> spans,
-                    CancellationToken cancellationToken)
-                {
-                    throw new NotImplementedException();
-                }
-            }
-
-            private class TestRazorDocumentPropertiesService : IRazorDocumentPropertiesService
-            {
-                public bool DesignTimeOnly => false;
-
-                public string DiagnosticsLspClientName => "RazorCSharp";
-            }
-        }
 
         [Obsolete("Use " + nameof(LSPProjectSnapshotManagerDispatcher))]
         private class TestProjectSnapshotManagerDispatcher : ProjectSnapshotManagerDispatcher
